Debug Payment API stripe -  Pass tempUSerID for testing -  booking successfull - stripe payment gateway working -  payment details entered in DB awaiting status.

+++ DataAccess/Data/RoomOrderDetails.cs
@@ -30,7 +30,7 @@ namespace DataAccess.Data
         public DateTime ActualCheckOutDate { get; set; }
 
         [Required]
-        public long TotalCost { get; set; }
+        public double TotalCost { get; set; }
 
         [Required]
         public int RoomId { get; set; }
